science/model/pyproject.toml
[project]
name = "science-model"
version = "0.1.0"
description = "Shared data models for Science research framework"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.0",
  "pyyaml>=6.0.3",
]

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/science_model"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 120

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
venvPath = "."
venv = ".venv"

[dependency-groups]
dev = [
    "jsonschema>=4.26.0",
    "pytest>=9.0",
    "ruff>=0.15.7",
]
